pyproject.toml
#:schema https://json.schemastore.org/pyproject.json

[project]
name = "sistema_ordens_servicos"
version = "0.11.1"
description = "SISTEMA DE ORDENS DE SERVIÇOS"
authors = [
    {name = "Gabriel Bitencourt Marin", email = "gabrielbitercourt@gmail.com"},
]
dependencies = [
    "Flask-SQLAlchemy~=3.1",
    "python-dotenv~=1.0",
    "Flask-WTF~=1.2",
    "WTForms-SQLAlchemy~=0.4",
    "sqlmodel~=0.0",
    "pydantic[email]~=2.9",
    "pydantic-extra-types~=2.9",
    "phonenumbers~=8.13",
    "Flask-Alembic~=3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pyprojectx]
main = ["pdm"]

[tool.pyprojectx.aliases]
install = "@pdm install"
run = "@pdm run"
start = "@pdm start"
update = ["@pdm update", "@run pre-commit autoupdate"]

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
lint = [
    "pyright~=1.1",
    "ruff~=0.6",
    "djlint~=1.35",
]
commit = [
    "pre-commit~=4.0",
    "commitizen~=3.29",
]

[tool.pdm.scripts]
post_install = "pre-commit install"
start = "flask run"

[tool.pyright]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
ignore = ["pw"]

[tool.ruff]
fix = true
show-fixes = true
line-length = 79
extend-exclude = ["pw"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
    "ANN101",
    "ANN102",
    "D203",
    "D213",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "SLF001", "S105", "S106"]

[tool.ruff.lint.isort]
force-single-line = false
force-wrap-aliases = false
lines-after-imports = -1
lines-between-types = 0
split-on-trailing-comma = true

[tool.ruff.lint.pycodestyle]
max-doc-length = 72
max-line-length = 99

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
